+++ src/components/Snackbar/LbSnackbar.vue
@@ -195,7 +195,7 @@ defineOptions({
 .snackbar-message
   flex: 1 1 auto
   font-size: typography.$font-size-label-base
-  font-weight: typography.$font-weight-regular
+  font-weight: var(--lb-font-weight-body)
   line-height: typography.$line-height-normal
   color: var(--lb-text-neutral-contrast-high)
   word-wrap: break-word